Extract trace builder from log_results.py into reusable module
Move the hierarchical MLflow trace builder (~790 lines) from
log_results.py into agent_eval/mlflow/trace_builder.py:

- build_trace() — builds trace dict from stdout.log + run_result
- log_trace() — submits trace dict to MLflow
- make_span(), iso_to_ns(), summarize_tool_input() — helpers

Parameterized: trace_name (was hardcoded "rfe.speedrun") and
subagent_dir (was derived from stdout_path). log_results.py drops
from 999 to 206 lines, importing build_trace and log_trace.
